Implement Subscriptions module, add User role, and revamp UI theming
- Admin: Add a comprehensive Subscriptions module including plan management, customer assignment, history tracking, and today's delivery routes.
- New Role: Introduce the "User" role with a dedicated `UserMainActivity` and fragments for home, plans, and subscription status.
- Customer: Implement "Latest Order" dashboard cards and enforce a time-restricted ordering window (9:00 AM – 8:00 PM).
- UI/Theming: Refactor hardcoded colors across all layouts to use semantic theme attributes (`surface_background`, `textColorPrimary`) for robust Dark Mode support.
- Stock: Enhance stock updates with category-based grouping, headers, and improved input fields.
- Monthly Summary: Add WhatsApp sharing functionality and a custom month/year picker for sales summaries.
- Admin Items: Implement an item freeze/active toggle and update category sorting logic.
- Core: Centralize full-screen loading overlays, refresh layouts via `RefreshManager`, and standardized error handling with `NetworkMessageUtils`.
- Infrastructure: Update AGP to 8.7.2 and Kotlin to 2.0.21; integrate Glide with OkHttp for authenticated image loading.
- Auth: Refactor login and welcome screens with version display and improved loading animations.

+++ app/src/main/java/com/svd/svdagencies/data/model/admin/Items/AdminItem.kt
@@ -16,7 +16,7 @@ data class AdminItem(
     val selling_price: String?,
     val buying_price: String?,
     val mrp: String?,
-    val stock_quantity: Double?,
+    val stock_quantity: Int?,
     val pcs_count: Int?,
     val image: String?,
     val frozen: Boolean
